refactor: OSRUser 필드 명칭 변경 및 로그인 관련 클래스 정리
- OSRUser name 및 username 변경
- LoginRequest 클래스 생성
- UserConnection 위치 변경
