+++ api/src/main/java/com/opensourcereader/api/controller/auth/request/LoginRequest.java
@@ -1,0 +1,3 @@
+package com.opensourcereader.api.controller.auth.request;
+
+public record LoginRequest(String email, String password) {}
