+++ api/src/main/java/com/opensourcereader/api/security/DefaultUserDetailService.java
@@ -8,6 +8,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import com.opensourcereader.api.security.login.OSRUser;
+import com.opensourcereader.core.security.dto.UserConnection;
 import com.opensourcereader.core.user.entity.User;
 import com.opensourcereader.core.user.exception.UserNotFoundException;
 import com.opensourcereader.core.user.service.UserService;
@@ -21,9 +22,9 @@ public class DefaultUserDetailService implements UserDetailsService {
   private final UserService userService;
 
   @Override
-  public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+  public UserDetails loadUserByUsername(String nickname) throws UsernameNotFoundException {
     final User user =
-        userService.findByEmail(email).orElseThrow(() -> new UserNotFoundException(email));
+        userService.findByNickname(nickname).orElseThrow(() -> new UserNotFoundException(nickname));
 
     UserConnection userConnection = new UserConnection(user, Instant.now());
