+++ core/src/main/java/com/opensourcereader/core/security/repository/RefreshTokenRepository.java
@@ -10,4 +10,6 @@ import com.opensourcereader.core.user.entity.User;
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 
   Optional<RefreshToken> findByUser(User user);
+
+  void deleteRefreshTokenByToken(String token);
 }
